requirements.txt
aioconsole==0.3.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bidict==0.21.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
idna==3.2
loguru==0.5.3
multidict==5.1.0
pycodestyle==2.7.0
pymongo==3.12.0
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
win32-setctime==1.0.3
yarl==1.6.3
